# openapi.yaml
openapi: 3.0.3
info:
  title: ZTN_SIM IAMaaS API
  version: "1.0.0"
  description: |
    Versioned REST contract for IAMaaS tenants and core platform operations. The schema documents
    authentication, token issuance, trust-score inspection, policy management, audit/log retrieval,
    and supporting wallet/transaction flows. It is intended to be consumed by the auth-gateway,
    policy decision point, external HMS integrations, and telecom or mobile-money pilots.
servers:
  - url: https://localhost.localdomain.com
    description: Local HTTPS (Podman + nginx)
  - url: http://localhost.localdomain.com
    description: Local HTTP (if configured)
security:
  - BearerAuth: []
    ApiKeyAuth: []
paths:
...
  /api/v1/auth/tenant/roles:
    get:
      tags: [Tenant]
      summary: List tenant roles
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
    post:
      tags: [Tenant]
      summary: Create a tenant role
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
...
  /roles:
    get:
      tags: [Roles]
      summary: List all roles
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
  /user/{userId}:
    get:
      tags: [Roles]
      summary: Fetch roles for a specific user
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
  /assign_role:
    post:
      tags: [Roles]
      summary: Assign a role with access level to a user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
  responses:
...
    MessageResponse:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
  schemas:
...
    Message:
      type: object
      properties:
        message:
          type: string
...
    Role:
      type: object
      properties:
        id:
          type: integer
        role_name:
          type: string
        access_level:
          type: string
    RoleCreateRequest:
      type: object
      required: [role_name]
      properties:
        role_name:
          type: string
        access_level:
          type: string
...
    AssignRoleRequest:
      type: object
      required: [user_id, role_id]
      properties:
        user_id:
          type: integer
        role_id:
          type: integer
        access_level:
          type: string
...